Move request sending stuff into its own module
